Fix for Message Crossing

+++ Program1/ThreadedClient/src/edu/nku/common/LogicalClock.java
@@ -1,27 +1,26 @@
 package edu.nku.common;
 
 public class LogicalClock {
-	private double currentTime;
-	
-	public LogicalClock(){
+	private int currentTime;
+
+	public LogicalClock() {
 		currentTime = 0;
 	}
-	
-	public synchronized void incrementTimeBy1(){
+
+	public synchronized void incrementTimeBy1() {
 		currentTime++;
 	}
-	
-	public synchronized void incrementTimeBySleep(double sleepTime){
+
+	public synchronized void incrementTimeBySleep(int sleepTime) {
 		currentTime += sleepTime;
 	}
-	
-	public synchronized void decideIncrement(int otherTime){
+
+	public synchronized void decideIncrement(int otherTime) {
 		currentTime = Math.max(currentTime + 1, otherTime + 1);
 	}
-	
-	public synchronized double getCurrentTime(){
+
+	public synchronized int getCurrentTime() {
 		return currentTime;
 	}
-	
-	
+
 }
