+++ Program1/ThreadedClient/src/edu/nku/common/Message.java
@@ -4,9 +4,9 @@ public class Message {
 	public int fromThread;
 	public int toThread;
 	public String messageBody;
-	public double localTime;
+	public int localTime;
 	
-	public Message(int fromThread, int toThread, String messageBody, double localTime){
+	public Message(int fromThread, int toThread, String messageBody, int localTime){
 		this.fromThread = fromThread;
 		this.toThread = toThread;
 		this.messageBody = messageBody;
